remove paneutil
